refactor: 모델 패키지 안으로 이동
