+++ src/main/java/baseball/model/PlayType.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.model;
 
 public enum PlayType {
     STRIKE, BALL, NOTHING;
