+++ src/test/java/baseball/BallTest.java
@@ -1,5 +1,7 @@
 package baseball;
 
+import baseball.model.Ball;
+import baseball.model.PlayType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
